refactor paged attention custom op

+++ src/custom_ops/paged_attention_impl.cu
@@ -179,8 +179,8 @@ void reshape_and_cache(
     const void* key_cache,    // [num_blocks, block_size, num_heads, head_size]
     const void* value_cache,  // [num_blocks, block_size, num_heads, head_size]
     const int* slot_mapping,  // [num_tokens]
-    const int64_t* key_shapes,
-    const int64_t* value_shapes,
+    const int32_t* key_shapes,
+    const int32_t* value_shapes,
     const int64_t block_size) {
   int num_tokens = key_shapes[0];
   int num_heads = key_shapes[1];
